pyproject.toml
[project]
name = "os-chat"
version = "0.1.0"
description = "Chat Agent to talk to your OS"
authors = [
    {name = "Maximilian Wessendorf", email = "maxi-git@outlook.de"},
]
dependencies = [
    "openai==1.12.0",
    "osquery==3.0.7",
    "phidata==2.3.35",
    "prompt-toolkit==3.0.43",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.0.1",
]

[tool.pdm.scripts]
start = {call = "src.os_chat.main:main"}
benchmark = {call = "src.os_chat.benchmark:main"}
tests = "pytest tests/"
